*** empty log message ***

+++ objs/Wind.h
@@ -352,7 +352,7 @@ public:
 				int true_dir_bins, int delta_dir_bins);
 	int		AvgNambigVsCti(WindField* truth, float* avg_nambig,
                 float low_speed, float high_speed);
-    int     WvcVsCti(WindField* truth, unsigned int* count, float low_speed,
+    int     WvcVsCti(WindField* truth, float* count, float low_speed,
                 float high_speed);
 	int		RmsSpdErrVsCti(WindField* truth, float* rms_spd_err_array,
 				float* std_dev_array, float* std_err_array,
